Moved encodeMuxSDU and decodeMuxSDUHeader to its own module
They deserve to be in the same module, this makes easier to review that
one is a left inverse of the other.
