+++ network-mux/src/Network/Mux/Bearer/Queues.hs
@@ -20,8 +20,7 @@ import qualified Network.Mux as Mx
 import qualified Network.Mux.Interface as Mx
 import           Network.Mux.Types (MuxBearer)
 import qualified Network.Mux.Types as Mx
-import qualified Network.Mux.Egress as Mx
-import qualified Network.Mux.Ingress as Mx
+import qualified Network.Mux.Codec as Mx
 import           Network.Mux.Time as Mx
