+++ network-mux/src/Network/Mux/Bearer/Socket.hs
@@ -27,8 +27,7 @@ import qualified Network.Socket.ByteString.Lazy as Socket (recv, sendAll)
 import qualified Network.Mux as Mx
 import           Network.Mux.Types (MuxBearer)
 import qualified Network.Mux.Types as Mx
-import qualified Network.Mux.Egress as Mx
-import qualified Network.Mux.Ingress as Mx
+import qualified Network.Mux.Codec as Mx
 import qualified Network.Mux.Time as Mx
 
 hexDump :: BL.ByteString -> String -> IO ()
