+++ network-mux/src/Network/Mux/Ingress.hs
@@ -5,15 +5,11 @@
 {-# LANGUAGE ScopedTypeVariables   #-}
 
 module Network.Mux.Ingress (
-      decodeMuxSDUHeader
-    , demux
+      demux
     , ingressQueue
     ) where
 
 import           Control.Monad
-import qualified Data.Binary.Get as Bin
-import           Data.Bits
-import           Data.Word
 import qualified Data.ByteString.Lazy as BL
 import           Data.Array
 import           GHC.Stack
@@ -24,6 +20,7 @@ import           Control.Monad.Class.MonadSTM
 import           Control.Monad.Class.MonadThrow
 
 import           Network.Mux.Types
+import           Network.Mux.Codec
 
 negMiniProtocolMode :: MiniProtocolMode -> MiniProtocolMode
 negMiniProtocolMode ModeInitiator = ModeResponder
@@ -70,43 +67,10 @@ negMiniProtocolMode ModeResponder = ModeInitiator
 -- >                    | BlockFetch|
 -- >                    +-----------+
 
-{- | Decode a MuSDU header -}
-decodeMuxSDUHeader :: (HasCallStack , ProtocolEnum ptcl)
-                   => BL.ByteString -> Either MuxError (MuxSDU ptcl)
-decodeMuxSDUHeader buf =
-    case Bin.runGetOrFail dec buf of
-         Left  (_, _, e)  -> Left $ MuxError MuxDecodeError e callStack
-         Right (_, _, ph) ->
-             let mode  = getMode $ mshIdAndMode ph
-                 mid_m = getId $ mshIdAndMode ph .&. 0x7fff in
-             case mid_m of
-                  Left  e   -> Left  $ MuxError MuxUnknownMiniProtocol ("id = " ++ show e) callStack
-                  Right mid -> Right $ MuxSDU {
-                        msTimestamp = mshTimestamp ph
-                      , msId = mid
-                      , msMode = mode
-                      , msLength = mshLength ph
-                      , msBlob = BL.empty
-                      }
-  where
-    dec = do
-        ts <- Bin.getWord32be
-        mid <- Bin.getWord16be
-        len <- Bin.getWord16be
-        return $ MuxSDUHeader (RemoteClockModel ts) mid len
-
-    getMode mid =
-        if mid .&. 0x8000 == 0 then ModeInitiator
-                               else ModeResponder
-
-    getId :: ProtocolEnum ptcl => Word16 -> Either Word16 (MiniProtocolId ptcl)
-    getId n | Just ptcl <- toProtocolEnum n
-            = Right ptcl
-    getId a = Left a
-
--- | demux runs as a single separate thread and reads complete 'MuxSDU's from the underlying
--- Mux Bearer and forwards it to the matching ingress queue.
-demux :: (MonadSTM m, MonadSay m, MonadThrow (STM m), Ord ptcl, Enum ptcl, Show ptcl
+-- | demux runs as a single separate thread and reads complete 'MuxSDU's from
+-- the underlying Mux Bearer and forwards it to the matching ingress queue.
+demux :: (MonadSTM m, MonadSay m, MonadThrow (STM m), Ord ptcl, Enum ptcl, Show
+      ptcl
          , MiniProtocolLimits ptcl, HasCallStack)
       => PerMuxSharedState ptcl m -> m ()
 demux pmss = forever $ do
